requirements.txt
neo4j==5.19.0
flask==3.0.3
flask-cors==4.0.0
pydot==2.0.0
anthropic==0.25.0
requests==2.31.0
python-dotenv==1.0.1
